+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 3.5.0 (in progress)
 ================
-* [#433] (https://github.com/oshi/oshi/pull/433): performance improvements for getProcesses() on Linux - [@bildechinger](https://github.com/bildechinger).
+* [#433] (https://github.com/oshi/oshi/pull/433): Performance improvements for getProcesses() on Linux - [@bildechinger](https://github.com/bildechinger).
+* [#455] (https://github.com/oshi/oshi/pull/455): Open files/handles support - [@spyhunter99](https://github.com/spyhunter99).
 * Your contribution here.
 
 3.4.4 (10/15/2017)